Moved DokanStoreProducts abstract class to 'includes/StoreProductsSections/' from 'includes/Abstracts/' directory
